+++ src/infrastructure/scrapers/polygon_ps5_top25.rs
@@ -5,7 +5,7 @@ use scraper::Html;
 pub struct PolygonPS5Top25;
 
 impl WebsiteScraper for PolygonPS5Top25 {
-    fn extract_games(&self, document: &Html, selectors: &Selectors) -> Result<Vec<(String, i32)>> {
+    fn extract_games(&self, document: &Html, selectors: &Selectors) -> Result<Vec<(String, u64)>> {
         let mut games = Vec::new();
 
         let names: Vec<String> = document
@@ -14,7 +14,7 @@ impl WebsiteScraper for PolygonPS5Top25 {
             .collect();
 
         for (i, name) in names.into_iter().enumerate() {
-            games.push((name, (i + 1) as i32));
+            games.push((name, (i + 1) as u64));
         }
 
         Ok(games)
