+++ src/infrastructure/scrapers/rockpapershotgun.rs
@@ -5,7 +5,7 @@ use scraper::Html;
 pub struct RPSScraper;
 
 impl WebsiteScraper for RPSScraper {
-    fn extract_games(&self, document: &Html, selectors: &Selectors) -> Result<Vec<(String, i32)>> {
+    fn extract_games(&self, document: &Html, selectors: &Selectors) -> Result<Vec<(String, u64)>> {
         let mut games = Vec::new();
 
         let names: Vec<String> = document
@@ -13,11 +13,11 @@ impl WebsiteScraper for RPSScraper {
             .map(|el| el.text().collect::<String>().trim().to_string())
             .collect();
 
-        let ranks: Vec<i32> = document
+        let ranks: Vec<u64> = document
             .select(selectors.rank.as_ref().unwrap())
             .filter_map(|el| {
                 let rank_str = el.text().collect::<String>();
-                rank_str.trim().parse::<i32>().ok()
+                rank_str.trim().parse::<u64>().ok()
             })
             .collect();
